renamed anyvendor helpers
