+++ generate.go
@@ -4,7 +4,7 @@ import (
 	"github.com/solo-io/anyvendor/anyvendor"
 	"github.com/solo-io/go-utils/log"
 	"github.com/solo-io/solo-kit/pkg/code-generator/cmd"
-	"github.com/solo-io/solo-kit/pkg/protodep"
+	"github.com/solo-io/solo-kit/pkg/code-generator/sk_anyvendor"
 )
 
 //go:generate go run generate.go
@@ -20,17 +20,17 @@ func main() {
 		SkipGeneratedTests: true,
 		ProtoDepConfig: &anyvendor.Config{
 			Local: &anyvendor.Local{
-				Patterns: []string{"test/**/*.proto", "api/**/*.proto", protodep.SoloKitMatchPattern},
+				Patterns: []string{"test/**/*.proto", "api/**/*.proto", sk_anyvendor.SoloKitMatchPattern},
 			},
 			Imports: []*anyvendor.Import{
 				{
-					ImportType: &anyvendor.Import_GoMod{GoMod: protodep.ExtProtoMatcher},
+					ImportType: &anyvendor.Import_GoMod{GoMod: sk_anyvendor.ExtProtoMatcher},
 				},
 				{
-					ImportType: &anyvendor.Import_GoMod{GoMod: protodep.EnvoyValidateProtoMatcher},
+					ImportType: &anyvendor.Import_GoMod{GoMod: sk_anyvendor.EnvoyValidateProtoMatcher},
 				},
 				{
-					ImportType: &anyvendor.Import_GoMod{GoMod: protodep.GogoProtoMatcher},
+					ImportType: &anyvendor.Import_GoMod{GoMod: sk_anyvendor.GogoProtoMatcher},
 				},
 			},
 		},
