+++ pkg/code-generator/sk_anyvendor/gomod.go
@@ -1,4 +1,4 @@
-package protodep
+package sk_anyvendor
 
 import (
 	"github.com/solo-io/anyvendor/anyvendor"
